fix(login): serializable 불가로 detailsImpl 분리 #2
