+++ src/main/java/com/godsang/anytimedelivery/auth/details/UserDetailServiceImpl.java
@@ -1,0 +1,24 @@
+package com.godsang.anytimedelivery.auth.details;
+
+import com.godsang.anytimedelivery.auth.utils.UserAuthorityUtils;
+import com.godsang.anytimedelivery.user.entity.User;
+import com.godsang.anytimedelivery.user.repository.UserRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.stereotype.Component;
+
+@Component
+@RequiredArgsConstructor
+public class UserDetailServiceImpl implements UserDetailsService {
+    private final UserAuthorityUtils userAuthorityUtils;
+    private final UserRepository userRepository;
+
+    @Override
+    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
+        User user = userRepository.findByEmail(username)
+                .orElseThrow(() -> new UsernameNotFoundException("User not found."));
+        return new UserDetailsImpl(user, userAuthorityUtils.createAuthorities(user.getRole()));
+    }
+}
